refactor: extract MapPicker subcomponents
Extract SavedAddressChips, SaveAddressDialog, and ManualAddressEntry
from MapPicker to improve code health and maintainability.
